Refactor: Move config file generation and image folder validation to utilities module
- Moved logic for generating the config file to the utilities module for better organization.
- Updated the image folder handling to ensure the folder exists, improving code clarity and reusability.
